+++ tutorbranding/plugin.py
@@ -61,6 +61,7 @@ config = {
         # Customizations of the learner dashboard in Quince. May not apply if the MFE is redesigned.
         "HIDE_UPGRADE_BUTTON": True,
         "HIDE_DASHBOARD_SIDEBAR": False,
+        "HIDE_LOOKING_FOR_CHALLENGE_WIDGET": False,
         "HIDE_PROGRAMS": False,
         "FIT_COURSE_IMAGE": True,