+++ tutorbranding/__about__.py
@@ -1,4 +1,4 @@
 """
 Package version
 """
-__version__ = "17.5.0"
+__version__ = "17.6.0"